Import of ASNMTAP/ASNMTAP-3.000013 from CPAN.
gitpan-cpan-distribution: ASNMTAP
gitpan-cpan-version:      3.000013
gitpan-cpan-path:         ASNMTAP/ASNMTAP-3.000013.tar.gz
gitpan-cpan-author:       ASNMTAP
gitpan-cpan-maturity:     released

+++ applications/sbin/perfparse_crontab.sh
@@ -1,28 +1,36 @@
-#!/bin/ksh
+#!/bin/bash
 # ----------------------------------------------------------------------------------------------------------
-# © Copyright 2003-2006 Alex Peeters [alex.peeters@citap.be]
+# © Copyright 2003-2007 Alex Peeters [alex.peeters@citap.be]
 # ----------------------------------------------------------------------------------------------------------
-# 2006/xx/xx, v3.000.012, perfparse_crontab.sh
+# 2007/02/25, v3.000.013, perfparse_crontab.sh
 # ----------------------------------------------------------------------------------------------------------
 
-# export PATH=/usr/local/bin:/usr/local/sbin:/usr/sbin:/etc:/usr/ccs/bin:/usr/bin:/opt/csw/bin:/usr/ucb:/usr/local/mysql/bin:${PATH}
+if [ -f ~/.profile ]; then
+  source ~/.profile
+fi
 
-# if [ -d /opt/asnmtap/cpan-shared/lib/perl5 ]; then
-#   PERL5LIB=${PERL5LIB:+$PERL5LIB:}/opt/asnmtap/cpan-shared/lib/perl5
-#   MANPATH=${MANPATH:+$MANPATH:}/opt/asnmtap/cpan-shared/share/man
-#   export PERL5LIB MANPATH
-# fi
+if [ -f ~/.bash_profile ]; then
+  source ~/.bash_profile
+fi
 
-# export LD_LIBRARY_PATH=/opt/asnmtap/ssl/lib:/usr/local/lib/mysql:/usr/local/lib:/usr/lib:${LD_LIBRARY_PATH}
+logFilename="/opt/asnmtap/log/perfdata-asnmtap.log"
 
-umask 022
+if [ -e "$logFilename" ]; then
+  if [ -e "/usr/local/bin/date" ]; then
+    epoch=`/usr/local/bin/date '+%s'`
+  elif [ -e "/usr/bin/date" ]; then
+    epoch=`/usr/bin/date '+%s'`
+  elif [ -e "/bin/date" ]; then
+    epoch=`/bin/date '+%s'`
+  else
+    exit 1;
+  fi
 
-epoch=`/usr/local/bin/date '+%s'` 
-epochFilename="/opt/asnmtap/log/perfdata-asnmtap.log-$epoch"
+  epochFilename="$logFilename-$epoch"
 
-echo "$epochFilename";
+  echo "Filenames: '$logFilename', '$epochFilename'";
+  mv $logFilename $epochFilename
+  cat $epochFilename | /opt/asnmtap/perfparse/bin/perfparse-log2mysql
+  rm $epochFilename
+fi
 
-mv /opt/asnmtap/log/perfdata-asnmtap.log $epochFilename
-touch /opt/asnmtap/log/perfdata-asnmtap.log
-/usr/local/bin/cat $epochFilename | /opt/asnmtap/perfparse/bin/perfparse-log2mysql
-rm $epochFilename
